feat(web): match the lightbox gestures in the capture dialog
Reviewing a shot before upload and reviewing it after upload were two
different surfaces: the dialog could only change photos by tapping a 56px
thumb, and a two-finger touch on it did nothing.

Move the gesture plumbing into PhotoStage and render it in both, so a shot
swipes, pinches, double-taps and pans the same before and after it lands in
S3. The listeners now read the active photo and navigation from a ref rather
than through effect dependencies — as deps they tore down and re-attached all
six listeners on every photo change, on a surface the operator still had
fingers on. The zoom reset on a photo change moved from an effect to a
render-time adjustment so the fitted photo is what first paints.

Split the dialog chrome by mode. Shooting reserved height for a note field
and a confirm button that could not do anything yet, and the shutter row had
to hold thumbs, shutter and picker at once — two thumbs fit on a 390px
screen. Shooting now shows only the native-camera controls, with the count
on a last-shot square and the confirm compacted into the top chrome so the
fast path still uploads without a review step. Reviewing gets the photo
surface plus one action row, and the note collapses behind a button since it
is rarely filled in.
